LayoutTests/crypto/workers/subtle/hmac-generate-export-key-jwk-expected.txt

[Worker] Test exporting a HMAC key in workers with JWK format.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/hmac-generate-export-key-jwk.js
[Worker] Generating a key...
[Worker] Exporting a key...
PASS [Worker] key.kty is 'oct'
PASS [Worker] key.key_ops is ['sign', 'verify']
PASS [Worker] key.alg is 'HS1'
PASS [Worker] key.ext is true
PASS [Worker] Base64URL.parse(key.k).byteLength is 64
PASS successfullyParsed is true

TEST COMPLETE
